Added follow notification & fixed some page bugs
The database design of V2EX notification sys is complex and hard to
extend, which too many useless fields if I want to add another function
such as "Follow notification" it has not used some of fields.
The best solution to this situation, I think, is a text field with json
data, so the DB design will get more flexible for any function
extension.